fix(arm): build for 32-bit ARM, and add the CI job that would have caught it
`horus deploy --arch armv7` advertises 32-bit ARM and the installation guide
names Raspberry Pi 3, but horus_sys did not compile for
armv7-unknown-linux-gnueabihf at all:

    error[E0308]: mismatched types
       --> horus_sys/src/rt/linux.rs:317:17
        |
    306 | pub(super) fn get_memlock_limit() -> u64 {
        |                                      --- expected `u64` ...
    317 |                 rlim.rlim_cur
        |                 ^^^^^^^^^^^^^ expected `u64`, found `u32`

`libc::rlimit::rlim_cur` is `u64` on x86_64 and aarch64 but `u32` on armv7,
so returning the field directly compiled everywhere the CI actually built.
`u64::from` covers both widths.

The reason it survived is that nothing ever built for it — `grep -rn
'armv7\|gnueabihf' .github/workflows/` returned nothing across all fifteen
workflows, while aarch64 had a dedicated job. Adds `linux-armv7` alongside
`linux-arm64` and wires it into the `multi-platform-success` gate, so a
32-bit regression fails the build instead of reaching a robot.

The new job also syntax-checks `horus_cpp/include/horus/layout_contract.hpp`
with the 32-bit cross-compiler. Those static_asserts pin the size and field
offsets that the zero-copy C ABI reads Rust structs through; a divergence
there corrupts memory rather than failing loudly, and 32-bit ARM is where
alignment assumptions are most likely to differ. Verified passing locally on
both aarch64 and armv7.

Also fixes a false positive in the docs harness. `documented_env_vars_are_read`
reported `HORUS_API_KEY` as an unimplemented promise, but both doc sites name
it precisely to say it does not exist. Two gaps in the negation detector:
it matched "there is no" but not "there is deliberately no", and it scanned
one line at a time while the disclaimer routinely wraps across a line break.
Now strips adverbs and looks at the previous line too.

+++ horus_sys/src/rt/linux.rs
@@ -314,7 +314,11 @@ pub(super) fn get_memlock_limit() -> u64 {
             if rlim.rlim_cur == libc::RLIM_INFINITY {
                 u64::MAX
             } else {
-                rlim.rlim_cur
+                // `rlim_t` is 32-bit on armv7-unknown-linux-gnueabihf and
+                // 64-bit on x86_64/aarch64. `u64::from` covers both widths;
+                // returning the field directly fails to compile for 32-bit ARM,
+                // which `horus deploy --arch armv7` targets.
+                u64::from(rlim.rlim_cur)
             }
         } else {
             0
